合并代码，避免动态加载的不确定性
